+++ src/test/java/org/folio/entlinks/integration/internal/AuthoritySourceFilesServiceTest.java
@@ -28,42 +28,42 @@ class AuthoritySourceFilesServiceTest {
 
   @Test
   void fetchAuthoritySourceUrls_positive() {
-    var e1 = new AuthoritySourceFile(UUID.randomUUID(), "url1");
-    var e2 = new AuthoritySourceFile(UUID.randomUUID(), "url2");
+    var e1 = new AuthoritySourceFile(UUID.randomUUID(), "url1", "source-file-name-1");
+    var e2 = new AuthoritySourceFile(UUID.randomUUID(), "url2", "source-file-name-2");
     var sourceFiles = List.of(e1, e2);
 
     when(client.fetchAuthoritySourceFiles()).thenReturn(new AuthoritySourceFiles(sourceFiles));
 
-    var actual = service.fetchAuthoritySourceUrls();
+    var actual = service.fetchAuthoritySources();
 
     assertThat(actual)
       .hasSize(sourceFiles.size())
-      .contains(entry(e1.id(), e1.baseUrl()), entry(e2.id(), e2.baseUrl()));
+      .contains(entry(e1.id(), e1), entry(e2.id(), e2));
   }
 
   @Test
   void fetchAuthoritySourceUrls_positive_ignoreNullValues() {
-    var e1 = new AuthoritySourceFile(UUID.randomUUID(), "url1");
-    var e2 = new AuthoritySourceFile(null, "url2");
-    var e3 = new AuthoritySourceFile(UUID.randomUUID(), "url3");
-    var e4 = new AuthoritySourceFile(UUID.randomUUID(), null);
+    var e1 = new AuthoritySourceFile(UUID.randomUUID(), "url1", "source-file-name-1");
+    var e2 = new AuthoritySourceFile(null, "url2", "source-file-name-2");
+    var e3 = new AuthoritySourceFile(UUID.randomUUID(), "url3", "source-file-name-3");
+    var e4 = new AuthoritySourceFile(UUID.randomUUID(), null, null);
     var validSourceFiles = List.of(e1, e3);
     var sourceFiles = List.of(e1, e2, e3, e4);
 
     when(client.fetchAuthoritySourceFiles()).thenReturn(new AuthoritySourceFiles(sourceFiles));
 
-    var actual = service.fetchAuthoritySourceUrls();
+    var actual = service.fetchAuthoritySources();
 
     assertThat(actual)
       .hasSize(validSourceFiles.size())
-      .contains(entry(e1.id(), e1.baseUrl()), entry(e3.id(), e3.baseUrl()));
+      .contains(entry(e1.id(), e1), entry(e3.id(), e3));
   }
 
   @Test
   void fetchAuthoritySourceUrls_negative_emptyResponse() {
     when(client.fetchAuthoritySourceFiles()).thenReturn(new AuthoritySourceFiles(emptyList()));
 
-    assertThatThrownBy(() -> service.fetchAuthoritySourceUrls())
+    assertThatThrownBy(() -> service.fetchAuthoritySources())
       .isInstanceOf(FolioIntegrationException.class)
       .hasMessage("Authority source files are empty.");
   }
@@ -73,7 +73,7 @@ class AuthoritySourceFilesServiceTest {
     var cause = new IllegalArgumentException("test message");
     when(client.fetchAuthoritySourceFiles()).thenThrow(cause);
 
-    assertThatThrownBy(() -> service.fetchAuthoritySourceUrls())
+    assertThatThrownBy(() -> service.fetchAuthoritySources())
       .isInstanceOf(FolioIntegrationException.class)
       .hasCauseExactlyInstanceOf(cause.getClass())
       .hasMessage("Failed to fetch authority source files");